Test all from pixi task + only problem: FANS is run from global scope - if pixi global fans is there, that one is run....
