+++ test/pytest/test_homogenization_consistency.py
@@ -4,31 +4,6 @@ import json
 import pytest
 from fans_dashboard.core.utils import identify_hierarchy, extract_and_organize_data
 
-
-@pytest.fixture(
-    params=[
-        "test_J2Plasticity",
-        "test_LinearElastic",
-        "test_LinearThermal",
-        "test_PseudoPlastic",
-    ]
-)
-def test_files(request):
-    json_base_dir = os.path.join(
-        os.path.dirname(os.path.abspath(__file__)), "../input_files/"
-    )
-    h5_base_dir = os.path.join(
-        os.path.dirname(os.path.abspath(__file__)), "../../build/test/"
-    )
-
-    json_path = os.path.join(json_base_dir, f"{request.param}.json")
-    h5_path = os.path.join(h5_base_dir, f"{request.param}.h5")
-
-    if os.path.exists(json_path) and os.path.exists(h5_path):
-        return json_path, h5_path
-    pytest.skip(f"Required test files not found: {json_path} or {h5_path}")
-
-
 def test_homogenization_consistency(test_files):
     """
     This test verifies that the relationship stress_average = homogenized_tangent * strain_average
